backend/worker/piper-tts-worker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
numpy==1.26.4
piper-tts==1.3.0
